requirements.txt
streamlit
Pillow
google-genai